feat: initial table creation moved to the migration as well
