+++ src/packages/oab-vector-store/src/sqlite-migrations/store/002_add_session_and_expiry.ts
